+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 This changelog adheres to [Keep a CHANGELOG](http://keepachangelog.com/).
 
+## [Unreleased]
+###
+- [TT-3304] Handle no response in booking update API
+
 ## [3.6.0]
 ###
 - [ROT-114] Add drop off details